requirements.txt
python-telegram-bot==20.7
python-dotenv==0.21.0
